Fix fake inliner variables spilling
While generating state machine, we check whether the variable is fake by
checking the range of the variable BEFORE suspension point.
However, in some cases the variable ends BEFORE suspension point and
continues AFTER it. In this case, it is more safe to check the range
at the end of the suspension point.
 #KT-29317 Fixed